+++ packages/core/src/__tests__/gzip.spec.ts
@@ -20,8 +20,6 @@ const API_TEST_INPUT = JSON.stringify({
     {
       event: 'custom-event',
       distinct_id: 'user-distinct-id',
-      library: 'posthog-core-tests',
-      library_version: '2.0.0-alpha',
       properties: {
         $lib: 'posthog-core-tests',
         $lib_version: '2.0.0-alpha',
@@ -29,7 +27,6 @@ const API_TEST_INPUT = JSON.stringify({
       },
       timestamp: new Date().toISOString(),
       uuid: randomUUID(),
-      type: 'capture',
     },
   ],
   sent_at: new Date().toISOString(),